requirements.txt
bitsandbytes-cuda113==0.26.0
black==22.3.0
mypy==0.961
numpy==1.21.6
pandas==1.3.5
pre-commit==2.19.0
pylint==2.14.3
pytest==7.1.2
rich==12.4.4
scikit-learn==1.0.2
scipy==1.7.3
torch==1.11.0
transformers==4.20.0
wandb==0.12.18
